backend/requirements.txt
openai-whisper
google-cloud-texttospeech
fastapi
python-multipart
uvicorn
